+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog
 - `Navigate → Related Symbol...` for tests
 - Show file type icons in project tree view
 - more Live Templates for HTMLBars
+- Basic reference resolving for e.g. `DS.belongsTo('user')`
 - Generate Ember.js files via `ember generate`
 
 ![Blueprints Dialog](doc/blueprints-dialog.png)